+++ examples/02_ui.zig
@@ -186,7 +186,6 @@ pub fn main() !void {
         .size = .{ .width = 720, .height = 480 },
     });
     defer win.deinit();
-    win.partial_present = true;
 
     var demo = Demo{ .app = &app };
     demo.window = win;