+++ examples/04_components.zig
@@ -354,7 +354,6 @@ pub fn main() !void {
         .size = .{ .width = 720, .height = 720 },
     });
     defer win.deinit();
-    win.partial_present = true;
 
     var demo = Demo{ .app = &app };
     demo.window = win;